youtube-auto/requirements.txt
google-generativeai>=0.8.0
edge-tts>=6.1.9
requests>=2.31.0
moviepy>=1.0.3
Pillow>=10.3.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
numpy>=1.26.0
schedule>=1.2.0
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
google-genai>=1.0.0
